# backend/src/main/java/com/example/banking/dto/CustomerActivityResponse.java
package com.example.banking.dto;

import lombok.Builder;
import lombok.Data;

import java.util.List;

@Data
@Builder
public class CustomerActivityResponse {
    private List<AccountSummaryDto> accounts;
    private List<TransactionDto> items;
}