# backend/src/main/java/com/example/banking/dto/TransferRequest.java
package com.example.banking.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;
import lombok.Data;

import java.math.BigDecimal;
import java.util.UUID;

@Data
public class TransferRequest {
    @NotNull
    private UUID sourceAccountId;

    @NotBlank
    private String recipientName;

    @NotBlank
    private String recipientAccountNumber;

    @NotNull
    @Positive
    private BigDecimal amount;

    @NotBlank
    private String description;
}